// client/app/authenticate/companies/company.ts
export interface Company {
    _id?: string;
    company?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    password?: string;
    phone?: string;
    country?: string;
    state?: string;
    city?: string;
    address1?: string;
    address2?: string;
    postalCode?: string;
    role?: string;
    code?: string;
}